src/06-agents-coding-lg/requirements.txt
openai==1.45.1
pydantic==2.9.1
python-dotenv==1.0.1
tiktoken==0.7.0
azure-search-documents==11.5.1
azure-identity==1.17.1
streamlit==1.38.0
promptflow==1.15.0
promptflow-core==1.15.0
promptflow-azure==1.15.0
promptflow-tracing==1.15.0
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langchain-qdrant==0.1.4